refactor signup form
